Move sparse_isometry.py
