+++ python/src/qdk_chemistry/algorithms/state_preparation/__init__.py
@@ -8,6 +8,9 @@ quantum states from classical wavefunctions.
 # Licensed under the MIT License. See LICENSE.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
+from qdk_chemistry.algorithms.state_preparation.sparse_isometry import (
+    SparseIsometryGF2XStatePreparation,
+)
 from qdk_chemistry.algorithms.state_preparation.state_preparation import (
     StatePreparation,
     StatePreparationFactory,
@@ -15,6 +18,7 @@ from qdk_chemistry.algorithms.state_preparation.state_preparation import (
 )
 
 __all__ = [
+    "SparseIsometryGF2XStatePreparation",
     "StatePreparation",
     "StatePreparationFactory",
     "StatePreparationSettings",
