+++ python/src/qdk_chemistry/algorithms/state_preparation/sparse_isometry.py
@@ -43,8 +43,7 @@ from qiskit.compiler import transpile
 from qiskit.quantum_info import Statevector
 from qiskit.transpiler import PassManager
 
-from qdk_chemistry.algorithms import register
-from qdk_chemistry.algorithms.state_preparation import StatePreparation, StatePreparationSettings
+from qdk_chemistry.algorithms.state_preparation.state_preparation import StatePreparation, StatePreparationSettings
 from qdk_chemistry.data import Wavefunction
 from qdk_chemistry.utils.bitstring import bitstrings_to_binary_matrix, separate_alpha_beta_to_binary_string
 
@@ -853,6 +852,3 @@ def _is_diagonal_matrix(matrix: np.ndarray) -> bool:
 
     # All other cases: not safe for diagonal reduction
     return False
-
-
-register(lambda: SparseIsometryGF2XStatePreparation())
